+++ src/index.js
@@ -1,15 +1,23 @@
-import apnumber from './apnumber'
-import apstate from './apstate'
-import intcomma from './intcomma'
-import intword from './intword'
-import ordinal from './ordinal'
-import postal from './postal'
+import apdate from './apdate';
+import apmonth from './apmonth';
+import apnumber from './apnumber';
+import aptime from './aptime';
+import capfirst from './capfirst';
+import intcomma from './intcomma';
+import intword from './intword';
+import ordinal from './ordinal';
+import pluralize from './pluralize';
+import yesno from './yesno';
 
 export {
+  apdate,
+  apmonth,
   apnumber,
-  apstate,
+  aptime,
+  capfirst,
   intcomma,
   intword,
   ordinal,
-  postal
-}
+  pluralize,
+  yesno,
+};