+++ src/ordinal.js
@@ -1,13 +1,18 @@
-import isInteger from 'lodash/isInteger'
-import isNil from 'lodash/isNil'
-import includes from 'lodash/includes'
+import { isInteger, isNil } from './utils';
 
 /**
  * A list of suffixes for conversions.
  * @private
  * @type {Array}
  */
-var SUFFIXES = ['th', 'st', 'nd', 'rd', 'th', 'th', 'th', 'th', 'th', 'th']
+const SUFFIXES = ['th', 'st', 'nd', 'rd', 'th', 'th', 'th', 'th', 'th', 'th'];
+
+/**
+ * A list of funky English ordinals.
+ * @private
+ * @type {Array}
+ */
+const ENGLISH_ORDINAL_EXCEPTIONS = [11, 12, 13];
 
 /**
  * Converts an integer into its ordinal form. Handles the special cases of 11,
@@ -18,28 +23,29 @@ var SUFFIXES = ['th', 'st', 'nd', 'rd', 'th', 'th', 'th', 'th', 'th', 'th']
  * @return {String}
  * @example
  *
- * var journalize = require('journalize')
+ * var journalize = require('journalize');
  *
- * journalize.ordinal(5)
+ * journalize.ordinal(5);
  * // returns '5th'
  *
- * journalize.ordinal(13)
+ * journalize.ordinal(13);
  * // returns '13th'
  *
- * journalize.ordinal(103)
+ * journalize.ordinal(103);
  * // returns '103rd'
  */
-export default function ordinal (val) {
+export default function ordinal(val) {
   // if `val` is undefined or null, return an empty string
-  if (isNil(val)) return ''
+  if (isNil(val)) return '';
 
-  var convertedVal = +val
+  const convertedVal = +val;
 
   // if `convertedVal` is not an integer, return `val`
-  if (!isInteger(convertedVal)) return val
+  if (!isInteger(convertedVal)) return val;
 
   // if `convertedVal` is 11, 12 or 13, English gets weird
-  if (includes([11, 12, 13], convertedVal % 100)) return convertedVal + SUFFIXES[0]
+  if (ENGLISH_ORDINAL_EXCEPTIONS.indexOf(convertedVal % 100) > -1)
+    return convertedVal + SUFFIXES[0];
 
-  return convertedVal + SUFFIXES[convertedVal % 10]
+  return convertedVal + SUFFIXES[convertedVal % 10];
 }